Version 1.2
Added support for bcc and multiple file attachments. Fix for lines ending in 0x0A not being handled properly. I've been told this does not work in 6.0 and that it should be a "plugin". Till I make the leap to 6.0 and figure out what that means, anyone wishing to lend a hand on the conversion is most welcome to do so :)